the_ape/requirements.txt
Django<2.0
mysqlclient
requests
django_extensions<2.0
djangorestframework<3.0.6
django_filter<0.9.3
django-debug-toolbar==1.2.1
django-grappelli
django-model-utils

Pillow
coverage==3.7.1
Sphinx==1.2.2